moved setup_g into journalist_app/__init__.py
