+++ securedrop/journalist.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from datetime import datetime, timedelta
+from datetime import datetime
 import functools
 
 from flask import (request, render_template, send_file, redirect, flash,
@@ -34,32 +34,6 @@ def shutdown_session(exception=None):
     """Automatically remove database sessions at the end of the request, or
     when the application shuts down"""
     db_session.remove()
-
-
-@app.before_request
-def setup_g():
-    """Store commonly used values in Flask's special g object"""
-    if 'expires' in session and datetime.utcnow() >= session['expires']:
-        session.clear()
-        flash(gettext('You have been logged out due to inactivity'), 'error')
-
-    session['expires'] = datetime.utcnow() + \
-        timedelta(minutes=getattr(config, 'SESSION_EXPIRATION_MINUTES', 30))
-
-    uid = session.get('uid', None)
-    if uid:
-        g.user = Journalist.query.get(uid)
-
-    g.locale = i18n.get_locale()
-    g.text_direction = i18n.get_text_direction(g.locale)
-    g.html_lang = i18n.locale_to_rfc_5646(g.locale)
-    g.locales = i18n.get_locale2name()
-
-    if request.method == 'POST':
-        filesystem_id = request.form.get('filesystem_id')
-        if filesystem_id:
-            g.filesystem_id = filesystem_id
-            g.source = get_source(filesystem_id)
 
 
 def login_required(func):
